Added Government Ideologies
Added the file GovernmentIdeologies.js that creates an instance of the Binary Tree structure. This tree holds all ideologies that involve government control and personal freedom. Updated the calculate button to show on the screen whichever ideology is the closest to the current value of the slider. Also added the most extreme ideologies to the tree to start with.